Add datetime filter to banner
